Create MINI PROJECT - 1.java
In this mini project ,that generates a random number between 1 and 100 and prompts the user to guess the number. The program should provide feedback to the user indicating whether their guess is too large or too small. The program should continue prompting the user for guesses until they correctly guess the number. Finally, the program should reveal the correct number.